Send picture: move the ImageResizeRequestHandler.Resize to a new service
