+++ LetsTalk.Server.ImageProcessing.Abstractions/IImageResizeService.cs
@@ -1,0 +1,6 @@
+﻿namespace LetsTalk.Server.ImageProcessing.Abstractions;
+
+public interface IImageResizeService
+{
+    byte[] Resize(byte[] data, int maxWidth, int maxHeight);
+}
