+++ LetsTalk.Server.ImageProcessor/ImageProcessorServiceRegistration.cs
@@ -7,6 +7,7 @@ using KafkaFlow.Serializer;
 using KafkaFlow.TypedHandler;
 using LetsTalk.Server.Configuration.Models;
 using LetsTalk.Server.FileStorage.Utility;
+using LetsTalk.Server.ImageProcessing;
 
 namespace LetsTalk.Server.ImageProcessor;
 
@@ -19,6 +20,7 @@ public static class ImageProcessorServiceRegistration
         var kafkaSettings = KafkaSettingsHelper.GetKafkaSettings(configuration);
         services.AddPersistenceServices(configuration);
         services.AddFileStorageUtilityServices();
+        services.AddImageProcessingServices();
         services.AddKafka(
             kafka => kafka
                 .UseConsoleLog()
